Adding Power Interface (V0.1)
Broke the power input out from the backplane so that we don't need to embed the power conditioning into the backplane and make a cleaner airflow in the whole thing. This will allow for ITX supply input and similar without having power conditioning or to have specific power input boards with a matching power conditioning.